# jps-next/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Just Praise Someone API
  version: 1.0.0
  description: Just Praise SomeoneのAPI仕様書
servers:
  - url: /api
    description: APIのベースURL
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Praise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        givenUserId:
          type: string
          format: uuid
        receivedUserId:
          type: string
          format: uuid
        content:
          type: string
          maxLength: 255
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        sortOrder:
          type: integer
...
paths:
...
  /praises:
    get:
      summary: 褒めの一覧取得
      tags:
        - Praises
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [received, given]
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 褒めの一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Praise'
        '400':
          description: バリデーションエラー
    post:
      summary: 新しい褒めの作成
      tags:
        - Praises
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - receivedUserId
              properties:
                content:
                  type: string
                  maxLength: 255
                  description: 褒めの内容
                receivedUserId:
                  type: string
                  format: uuid
                  description: 褒められるユーザーのID
                skills:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: 関連するスキルのIDリスト
      responses:
        '201':
          description: 褒めの作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Praise'
...